Error handling for categories in use by a todo

+++ src/main/java/com/shirdheen/todos/to_dos_app_project/services/CategoryService.java
@@ -3,6 +3,7 @@ package com.shirdheen.todos.to_dos_app_project.services;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.shirdheen.todos.to_dos_app_project.dto.DTOMapper;
@@ -10,13 +11,16 @@ import com.shirdheen.todos.to_dos_app_project.dto.category.CategoryDTO;
 import com.shirdheen.todos.to_dos_app_project.dto.category.CategoryRequestDTO;
 import com.shirdheen.todos.to_dos_app_project.entities.Category;
 import com.shirdheen.todos.to_dos_app_project.repositories.CategoryRepository;
+import com.shirdheen.todos.to_dos_app_project.repositories.ToDoRepository;
 
 @Service
 public class CategoryService {
     private final CategoryRepository categoryRepository;
+    private final ToDoRepository todoRepository;
 
-    public CategoryService(CategoryRepository categoryRepository) {
+    public CategoryService(CategoryRepository categoryRepository, ToDoRepository todoRepository) {
         this.categoryRepository = categoryRepository;
+        this.todoRepository = todoRepository;
     }
 
     public List<CategoryDTO> getAllCategories() {
@@ -44,6 +48,11 @@ public class CategoryService {
     public void deleteCategory(Long id) {
         Category category = categoryRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Sorry, category not found"));
+        
+        boolean isCategoryInUse = todoRepository.existsByCategoryId(id);
+        if (isCategoryInUse) {
+            throw new DataIntegrityViolationException("Cannot delete cateory. It is assigned to existing Todos.");
+        }
 
         categoryRepository.delete(category);
     }
